dockers/encodocker/requirements.txt
ffmpeg-python
flask
connexion
requests
flask_restful
flask_jsonpify
pika